VMI tool to interact with the repair_driver kernel module.  -Ability to pass commands to the ring buffer and receive results of command execution Still pending: insertion of breakpoints.

+++ asm/moti/repair_driver.h
@@ -56,13 +56,13 @@ struct submod_table {
 
 /* structure for the ring channel */
 struct cmd_ring_channel {
-    unsigned long     cons;                   /* Next item to be consumed. */
-    unsigned long     prod;                   /* Next item to be produced. */
-    unsigned long     payload_buffer_size;    /* size_in_pages * page_size */
-    unsigned long     buf_order;
+    unsigned int     cons;                   /* Next item to be consumed. */
+    unsigned int     prod;                   /* Next item to be produced. */
+    unsigned int     payload_buffer_size;    /* size_in_pages * page_size */
+    unsigned int     buf_order;
     char              *recs;                  /* pointer to buffer data areas */
-    unsigned long     size_of_a_rec;          /* size of a single record */
-    unsigned long     size_in_recs;           /* size of the buffer in recs */
+    unsigned int     size_of_a_rec;          /* size of a single record */
+    unsigned int     size_in_recs;           /* size of the buffer in recs */
 };
